+++ examples/simple-crate/crates.nix
@@ -8,8 +8,7 @@
     crateName = "my-crate";
   in {
     # declare projects
-    # TODO: change this to your crate's path
-    nci.projects.${crateName}.path = ./.;
+    nci.projects."simple".path = ./.;
     # configure crates
     nci.crates.${crateName} = {};
   };